+++ src/components/common/button/Button.module.css
@@ -15,7 +15,6 @@
   height: 32px;
   font-size: 14px;
   font-weight: 500;
-  font-family: 'pretendard';
 }
 
 .button-content {
@@ -97,7 +96,6 @@
   height: 70px;
   font-size: 16px;
   font-weight: 600;
-  font-family: 'pretendard';
 }
 
 .dashboard.add .button-content {
