+++ src/components/product/mydashboard/InvitationList.module.css
@@ -1,40 +1,3 @@
-.dashboard-list {
-  max-width: 1022px;
-  padding-top: 40px;
-  margin-left: 340px;
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  gap: 13px;
-}
-
-.color-circle {
-  width: 8px;
-  height: 8px;
-  border-radius: 50%;
-  margin-right: 16px;
-  display: inline-block;
-}
-
-.crown-icon {
-  margin-left: 8px;
-  display: flex;
-  align-items: center;
-}
-
-.pagination {
-  margin-left: 340px;
-  margin-top: 12px;
-  display: flex;
-  align-items: center;
-  gap: 16px;
-}
-
-.pagination-text {
-  font-size: 14px;
-  font-weight: 400;
-  color: var(--black-medium);
-}
-
 .head {
   padding: 32px 28px;
   margin-top: 72px;
@@ -76,7 +39,6 @@
   margin-left: 340px;
   background-color: var(--white);
   width: 1022px;
-  height: 650px;
 }
 
 .table {
@@ -119,34 +81,15 @@
 }
 
 @media screen and (max-width: 1199px) {
-  .dashboard-list {
-    max-width: 504px;
-    margin-left: 200px;
-    grid-template-columns: repeat(2, 1fr);
-  }
-
   .invitation-container {
     max-width: 504px;
-  }
-
-  .pagination,
-  .invitation-container {
     margin-left: 200px;
   }
 }
 
 @media screen and (max-width: 743px) {
-  .dashboard-list {
-    margin-left: 90px;
-    grid-template-columns: repeat(1, 1fr);
-  }
-
   .invitation-container {
     max-width: 260px;
-  }
-
-  .pagination,
-  .invitation-container {
     margin-left: 90px;
   }
 }
