+++ src/pages/_app.tsx
@@ -1,5 +1,6 @@
 import '@/styles/globals.css';
 import '@/styles/fonts.css';
+import '@/styles/normalize.css';
 import type { AppProps } from 'next/app';
 
 export default function App({ Component, pageProps }: AppProps) {
